pyproject.toml
[project]
name = "fastapi-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
